requirements.txt
# Python version requirement
# Tested with Python==3.11.x

blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3

# Environment variables handling
python-dotenv==1.0.1

# Day 18 dependencies — PDF generation and chart rendering
fpdf2==2.7.9
matplotlib==3.9.2
requests==2.32.3

# Day 19 dependencies — Dashboard, API, AJAX filtering
# jQuery is served via CDN in HTML (no pip install needed)
pandas==2.2.3
python-dateutil==2.9.0.post0

# Day 20 dependencies — PDF report automation & email sending
# smtplib and email are included in Python stdlib (no pip install required)
# cron jobs handled at system level, no pip install needed

# Day 21 dependencies — Interactive charts & enhanced reports
# Chart.js is served via CDN in HTML (no pip install needed)
# No new Python packages required — all interactivity handled in JS

# Day 22 dependencies — Dark Mode toggle & dynamic chart themes
# DejaVu fonts are preinstalled on most Linux distros, otherwise install via OS package manager:
#   Debian/Ubuntu: sudo apt install fonts-dejavu-core
# For emoji support in PDFs (optional), install Noto Color Emoji font:
#   Debian/Ubuntu: sudo apt install fonts-noto-color-emoji
# No additional pip packages required for Day 22 features

# Day 23 dependencies — OSINT phishing pipeline
beautifulsoup4==4.12.3
googlesearch-python==1.2.5
sqlite-utils==3.36

# Day 24 dependencies — Bug fixes & performance optimization
pytest==8.3.3
tqdm==4.67.1
pysqlite3-binary==0.5.4
